Seperate EventController & AdminEventController
