+++ app/Resources/views/admin/event/_partial/card.html.twig
@@ -27,14 +27,14 @@
         {% if from_admin %}
             {% if event.needProxy %}
                 <div class="left">
-                    <a href="{{ path("event_proxies_list", {'id': event.id}) }}"><i class="material-icons left">list</i>Procurations</a>
+                    <a href="{{ path("admin_event_proxies_list", {'id': event.id}) }}"><i class="material-icons left">list</i>Procurations</a>
                 </div>
                 <div class="left">
-                    <a href="{{ path("event_signatures", {'id': event.id}) }}"><i class="material-icons left">print</i>Emargement</a>
+                    <a href="{{ path("admin_event_signatures", {'id': event.id}) }}"><i class="material-icons left">print</i>Emargement</a>
                 </div>
             {% endif %}
             <div class="right">
-                <a href="{{ path("event_edit", {'id': event.id}) }}"><i class="material-icons left">edit</i>Editer</a>
+                <a href="{{ path("admin_event_edit", {'id': event.id}) }}"><i class="material-icons left">edit</i>Editer</a>
             </div>
         {% else %}
             {% if event.needProxy %}
