+++ app/Resources/views/admin/event/edit.html.twig
@@ -5,7 +5,7 @@
 {% block breadcrumbs %}
 <a href="{{ path('homepage') }}"><i class="material-icons">home</i></a><i class="material-icons">chevron_right</i>
 <a href="{{ path('admin') }}"><i class="material-icons">build</i>&nbsp;Administration</a><i class="material-icons">chevron_right</i>
-<a href="{{ path('event_index') }}"><i class="material-icons">event</i>&nbsp;Événements</a><i class="material-icons">chevron_right</i>
+<a href="{{ path('admin_event_index') }}"><i class="material-icons">event</i>&nbsp;Événements</a><i class="material-icons">chevron_right</i>
 <i class="material-icons">edit</i>&nbsp;Editer
 {% endblock %}
 
@@ -89,10 +89,10 @@
     {% if event.needProxy %}
         <ul>
             <li>
-                <a href="{{ path("event_proxies_list",{'id':event.id}) }}"><i class="material-icons">list</i>Procurations</a>
+                <a href="{{ path("admin_event_proxies_list",{'id':event.id}) }}"><i class="material-icons">list</i>Procurations</a>
             </li>
             <li>
-                <a href="{{ path("event_signatures",{'id':event.id}) }}"><i class="material-icons">list</i>Emargement</a>
+                <a href="{{ path("admin_event_signatures",{'id':event.id}) }}"><i class="material-icons">list</i>Emargement</a>
             </li>
         </ul>
     {% endif %}
