+++ app/Resources/views/admin/event/index.html.twig
@@ -25,10 +25,10 @@
 
 {% if is_granted("ROLE_ADMIN") %}
     <br />
-    <a href="{{ path("event_list") }}" class="btn"><i class="material-icons left">list</i>Tous les événements</a>
-    <a href="{{ path("proxies_list") }}" class="btn"><i class="material-icons left">list</i>Toutes les procurations</a>
+    <a href="{{ path("admin_event_list") }}" class="btn"><i class="material-icons left">list</i>Tous les événements</a>
+    <a href="{{ path("admin_proxies_list") }}" class="btn"><i class="material-icons left">list</i>Toutes les procurations</a>
     <br />
-    <a href="{{ path('event_new') }}" class="btn"><i class="material-icons left">add</i>Ajouter un événement</a>
+    <a href="{{ path('admin_event_new') }}" class="btn"><i class="material-icons left">add</i>Ajouter un événement</a>
 {% endif %}
 <a href="{{ path('event_widget_generator') }}" class="btn"><i class="material-icons left">tune</i>Générer un widget</a>
 {% endblock %}
