+++ app/Resources/views/default/index.html.twig
@@ -162,7 +162,7 @@
 
                 <br />
                 <br />
-                <a href="{{ path("event_public_list") }}" class="btn blue" style="position: relative">
+                <a href="{{ path("event_index") }}" class="btn blue" style="position: relative">
                     <i class="material-icons left" >event</i>Les événements à venir
                     <span class="bubble_counter red white-text">{{ eventsFuture | length }}</span>
                 </a>
