+++ app/Resources/views/event/_partial/widget.html.twig
